updating workflow so that wall-e docker image is built in the same stage as deployment to PROD

+++ CI/server_scripts/build_wall_e/deploy_to_prod_discord_guild.sh
@@ -5,18 +5,38 @@
 set -e -o xtrace
 # https://stackoverflow.com/a/5750463/7734535
 
+
+if [ -z "${DOCKER_HUB_PASSWORD}" ]; then
+    echo "DOCKER_HUB_PASSWORD is not set"
+    exit 1
+fi
+
+
+if [ -z "${DOCKER_HUB_USER_NAME}" ]; then
+    echo "DOCKER_HUB_USER_NAME is not set"
+    exit 1
+fi
+
 rm ${DISCORD_NOTIFICATION_MESSAGE_FILE} || true
 
-export prod_container_name="${COMPOSE_PROJECT_NAME}_wall_e"
-export prod_website_container_name="${COMPOSE_PROJECT_NAME}_leveling_website"
-export prod_container_db_name="${COMPOSE_PROJECT_NAME}_wall_e_db"
-export docker_compose_file="CI/server_scripts/build_wall_e/docker-compose.yml"
 export compose_project_name=$(echo "$COMPOSE_PROJECT_NAME" | awk '{print tolower($0)}')
-export COMPOSE_PROJECT_NAME_lower=$(echo "$COMPOSE_PROJECT_NAME" | awk '{print tolower($0)}')
-export prod_image_name_lower_case=$(echo "$prod_container_name" | awk '{print tolower($0)}')
-export prod_website_image_name_lower_case=$(echo "$prod_website_container_name" | awk '{print tolower($0)}')
-export ORIGIN_IMAGE="sfucsssorg/wall_e"
 
+export docker_compose_file="CI/server_scripts/build_wall_e/docker-compose.yml"
+
+export prod_container_name="${COMPOSE_PROJECT_NAME}_wall_e"
+export prod_image_name_lower_case=$(echo "$prod_container_name" | awk '{print tolower($0)}')
+
+export prod_website_container_name="${COMPOSE_PROJECT_NAME}_leveling_website"
+export prod_website_image_name_lower_case=$(echo "$prod_website_container_name" | awk '{print tolower($0)}')
+
+export prod_container_db_name="${COMPOSE_PROJECT_NAME}_wall_e_db"
+
+export docker_registry="sfucsssorg"
+export wall_e_top_base_image="wall_e_base"
+export ORIGIN_IMAGE="${docker_registry}/wall_e"
+export WALL_E_PYTHON_BASE_IMAGE="${docker_registry}/wall_e_python"
+
+export wall_e_top_base_image_dockerfile="CI/server_scripts/build_wall_e/Dockerfile.wall_e_base"
 
 docker rm -f ${prod_container_name} || true
 docker rm -f ${prod_website_container_name} || true
@@ -24,6 +44,23 @@ docker image rm -f ${prod_image_name_lower_case} || true
 docker image rm -f ${prod_website_image_name_lower_case} || true
 docker rm -f TEST_master_leveling_website|| true
 docker volume create --name="${COMPOSE_PROJECT_NAME}_logs"
+
+
+re_create_top_base_image () {
+    docker image rm -f "${prod_image_name_lower_case}" "${wall_e_top_base_image}" "${ORIGIN_IMAGE}"
+    docker system prune -f
+    docker build --no-cache -t ${wall_e_top_base_image} -f ${wall_e_top_base_image_dockerfile} \
+    --build-arg CONTAINER_HOME_DIR=${CONTAINER_HOME_DIR} \
+    --build-arg WALL_E_BASE_ORIGIN_NAME=${WALL_E_PYTHON_BASE_IMAGE} .
+    docker tag ${wall_e_top_base_image} ${ORIGIN_IMAGE}
+    echo "${DOCKER_HUB_PASSWORD}" | docker login --username=${DOCKER_HUB_USER_NAME} --password-stdin
+    docker push ${ORIGIN_IMAGE}
+    docker image rm -f "${prod_image_name_lower_case}" "${wall_e_top_base_image}" "${ORIGIN_IMAGE}"
+    docker system prune -f
+}
+
+re_create_top_base_image
+
 docker-compose -f "${docker_compose_file}" up -d
 sleep 20
